+++ src/app/app.component.scss
@@ -1,0 +1,5 @@
+h1{
+  font-family: Roboto, SansSerif;
+  font-weight: bolder;
+  color: aliceblue;
+}
