+++ src/app/app.component.ts
@@ -11,23 +11,10 @@ import * as StarWarsActions from '../state/star-wars/star-wars.actions';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  title = 'app';
-
-  films$: Observable<any>;
-  categoryContent: any;
-  films_content: any;
+  title = 'THE STAR WARS API APP';
 
   constructor(
-    private store: Store<State>
   ) {
-    this.films$ = this.store;
-    this.films$.subscribe((data) => {
-      this.films_content = data;
-    });
-    this.store.dispatch(
-      new StarWarsActions.GetFilmsAction({
-      })
-    );
   }
 
 }
