+++ src/app/app.module.ts
@@ -12,8 +12,11 @@ import { FilmsComponent } from './films/films.component';
 import { AppRoutingModule } from './app-routing.module';
 import { NavComponent } from './nav/nav.component';
 
-import  { StarWarsReducer } from '../state/star-wars/star-wars.reducer';
+import { StarWarsReducer } from '../state/star-wars/star-wars.reducer';
 import { StarWarsEffects } from '../state/star-wars/star-wars.effects';
+import { FilmsReducer } from '../state/films/films.reducer';
+import { FilmEffects } from '../state/films/films.effects'
+
 
 @NgModule({
   declarations: [
@@ -25,8 +28,8 @@ import { StarWarsEffects } from '../state/star-wars/star-wars.effects';
   imports: [
     BrowserModule,
     AppRoutingModule,
-    StoreModule.forRoot(({starWars: StarWarsReducer})),
-    EffectsModule.forRoot([StarWarsEffects]),
+    StoreModule.forRoot(({ film: FilmsReducer })),
+    EffectsModule.forRoot([ FilmEffects]),
     HttpModule,
     HttpClientModule
   ],
