+++ src/app/films/films.component.html
@@ -1,3 +1,24 @@
-<p>
-  films works!
-</p>
+<div *ngIf="films_content.loading; else elseBlock">
+  Loading ...
+</div>
+<ng-template #elseBlock>
+  <div class="sw-group" *ngIf="films_content.films_content && films_content.films_content.results ">
+    <ul *ngFor="let film of films_content.films_content.results">
+      <li >
+        Film: {{film.title}}
+      </li>
+      <li >
+        Episode ID: {{film.episode_id}}
+      </li>
+      <li >
+        Director: {{film.director}}
+      </li>
+      <li >
+        Release Data: {{film.release_date}}
+      </li>
+    </ul>
+
+  </div>
+</ng-template>
+
+
