+++ src/app/films/films.component.scss
@@ -1,0 +1,18 @@
+ul{
+  font-family: 'Roboto', 'SansSerif';
+  color: aliceblue;
+  background-color: black;
+  list-style-type: none;
+  margin: 3px;
+  border-radius: 4px;
+  padding: 15px;
+  width: 400px;
+}
+
+.sw-group{
+  display: flex;
+  max-width: 100%;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  margin: 0 50px;
+}
