+++ src/app/films/films.component.ts
@@ -1,4 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
+import { Store } from '@ngrx/store';
+
+import { FilmState  } from '../../state/films/films.state';
+import * as FilmActions from '../../state/films/films.actions';
 
 @Component({
   selector: 'app-films',
@@ -7,10 +12,24 @@ import { Component, OnInit } from '@angular/core';
 })
 export class FilmsComponent implements OnInit {
 
-  constructor() { }
+  films$: Observable<any>;
+  categoryContent: any;
+  films_content: any;
+  constructor( private store: Store<FilmState>) {
+    this.films$ = this.store;
+    this.films$.subscribe((data) => {
+      this.films_content = data.film;
+    });
+    this.store.dispatch(
+      new FilmActions.GetFilmsAction()
+    );
+  }
 
   ngOnInit() {
   }
 
 }
+
+
+
