+++ src/app/nav/nav.component.html
@@ -1,8 +1,9 @@
-<h1>Welcome to the Star Wars App</h1>
-<a routerLink="/landing">Landing</a>
-<a routerLink="/films">Films</a>
-<a routerLink="/people">People</a>
-<a routerLink="/starships">Starships</a>
-<a routerLink="/vehicles">Vehicles</a>
-<a routerLink="/species">Species</a>
-<a routerLink="/planets">Planets</a>
+<div>
+  <a routerLink="/landing">Landing</a>
+  <a routerLink="/films">Films</a>
+  <a routerLink="/people">People</a>
+  <a routerLink="/starships">Starships</a>
+  <a routerLink="/vehicles">Vehicles</a>
+  <a routerLink="/species">Species</a>
+  <a routerLink="/planets">Planets</a>
+</div>
