+++ src/app/nav/nav.component.scss
@@ -1,0 +1,13 @@
+a{
+  text-decoration: none;
+  font-size: 20px;
+  font-family: 'Roboto', 'SansSerif';
+  font-weight: bold;
+  color: lightblue;
+  padding: 30px;
+}
+
+div{
+  display: flex;
+  justify-content: space-around;
+}
